speech_recognition/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
moviepy==1.0.3
speechrecognition==3.10.0
psycopg2-binary==2.9.7